refactor to use base LoginCipherScene and LoginCipherRequest
